UAST: utilize identifier test base
