+++ plugins/uast-kotlin-base/build.gradle.kts
@@ -11,8 +11,16 @@ dependencies {
     // BEWARE: UAST should not depend on IJ platform so that it can work in Android Lint CLI mode (where IDE is not available)
     compileOnly(intellijCoreDep()) { includeJars("intellij-core", "asm-all", rootProject = rootProject) }
     compileOnly(intellijPluginDep("java")) { includeJars("java-api", "java-impl") }
+
+    testImplementation(commonDep("junit:junit"))
+    testCompileOnly(intellijPluginDep("java")) { includeJars("java-api", "java-impl") }
+    testImplementation(projectTests(":compiler:tests-common"))
+    testCompileOnly(intellijDep("ideaIC"))
 }
 
 sourceSets {
     "main" { projectDefault() }
+    "test" { projectDefault() }
 }
+
+testsJar ()
