+++ plugins/uast-kotlin-base/tests/org/jetbrains/uast/test/common/kotlin/IndentedPrintingVisitor.kt
@@ -1,12 +1,11 @@
 /*
- * Copyright 2010-2019 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 package org.jetbrains.uast.test.common.kotlin
 
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiElementVisitor
-import com.intellij.psi.PsiFile
 import org.jetbrains.uast.UFile
 import kotlin.reflect.KClass
 
@@ -36,9 +35,9 @@ abstract class IndentedPrintingVisitor(val shouldIndent: (PsiElement) -> Boolean
 
     val result: String
         get() = builder.toString()
-}
 
-fun IndentedPrintingVisitor.visitUFileAndGetResult(uFile: UFile): String {
-    (uFile.sourcePsi as PsiFile).accept(this)
-    return result
+    fun visitUFileAndGetResult(uFile: UFile): String {
+        uFile.sourcePsi.accept(this)
+        return result
+    }
 }
