+++ plugins/uast-kotlin/build.gradle.kts
@@ -44,6 +44,7 @@ dependencies {
     testCompile(project(":compiler:util"))
     testCompile(project(":compiler:cli"))
     testCompile(projectTests(":idea:idea-test-framework"))
+    testCompile(projectTests(":plugins:uast-kotlin-base"))
 
     testCompile(project(":idea:idea-native")) { isTransitive = false }
     testCompile(project(":idea:idea-gradle-native")) { isTransitive = false }
