+++ plugins/uast-kotlin/tests/AbstractKotlinIdentifiersTest.kt
@@ -1,14 +1,14 @@
+/*
+ * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
 package org.jetbrains.uast.test.kotlin
 
-import com.intellij.psi.PsiElement
-import org.jetbrains.uast.*
-import org.jetbrains.uast.test.common.UElementToParentMap
+import org.jetbrains.uast.UFile
 import org.jetbrains.uast.test.common.kotlin.IdentifiersTestBase
-import org.jetbrains.uast.test.common.visitUFileAndGetResult
+import org.jetbrains.uast.test.common.kotlin.asRefNames
 import org.jetbrains.uast.test.env.kotlin.assertEqualsToFile
 import java.io.File
-import kotlin.test.assertNotNull
-
 
 abstract class AbstractKotlinIdentifiersTest : AbstractKotlinUastTest(), IdentifiersTestBase {
 
@@ -22,16 +22,3 @@ abstract class AbstractKotlinIdentifiersTest : AbstractKotlinUastTest(), Identif
         assertEqualsToFile("refNames", getTestFile(testName, "refNames.txt"), file.asRefNames())
     }
 }
-
-private fun refNameRetriever(psiElement: PsiElement): UElement? =
-    when (val uElement = psiElement.toUElementOfExpectedTypes(UCallExpression::class.java, UReferenceExpression::class.java)) {
-        is UReferenceExpression -> uElement.referenceNameElement
-        is UCallExpression -> uElement.classReference?.referenceNameElement
-        else -> null
-    }?.also {
-        assertNotNull(it.sourcePsi, "referenceNameElement should have physical source, origin = $psiElement")
-    }
-
-fun UFile.asRefNames() = object : UElementToParentMap(::refNameRetriever) {
-    override fun renderSource(element: PsiElement): String = element.javaClass.simpleName
-}.visitUFileAndGetResult(this)
